+++ index.php
@@ -3,28 +3,18 @@
  * Home File
  *
  * @author      David Voglgsang
- * @version     1.0
+ * @version     1.1
  *
 */
 
 get_header();
-
-$obj = get_queried_object();
-// page options
-$options = prefix_template::PageOptions($obj->ID);
-// body class
-$pt = $obj ? 'pt-' . $obj->post_type : '';
 ?>
 <section id="home" <?php echo prefix_template::AddContainer(prefix_template::$template_container, true); ?>>
-  <?php if(!in_array('title', $options)): ?>
-    <h1><?php the_archive_title(); ?></h1>
-  <?php endif; ?>
+  <h1><?php the_archive_title(); ?></h1>
   <?php if (have_posts() ) : while (have_posts()) : the_post() ?>
     <article>
-      <?php if(!in_array('title', $options)): ?>
         <h2><?php the_title(); ?></h2>
-      <?php endif; ?>
-      <?php the_content(); ?>
+      <?php the_excerpt(); ?>
     </article>
     <?php get_sidebar(); ?>
   <?php endwhile; endif; ?>